Fix format warnings

+++ real_malloc/main.c
@@ -204,7 +204,7 @@ void run_challenge(const char *trace_file_name, size_t min_size,
         allocated += size;
         void *ptr = malloc_func(size);
         if (trace_fp) {
-          fprintf(trace_fp, "a %lld %ld\n", (uint64_t)ptr, size);
+          fprintf(trace_fp, "a %llu %ld\n", (unsigned long long)ptr, size);
         }
         memset(ptr, tag, size);
         object_t object = {ptr, size, tag};
@@ -236,7 +236,7 @@ void run_challenge(const char *trace_file_name, size_t min_size,
         }
         free_func(object.ptr);
         if (trace_fp) {
-          fprintf(trace_fp, "f %lld %ld\n", (uint64_t)object.ptr, object.size);
+          fprintf(trace_fp, "f %llu %ld\n", (unsigned long long)object.ptr, object.size);
         }
       }
 
@@ -340,7 +340,7 @@ void *mmap_from_system(size_t size) {
                    MAP_PRIVATE | MAP_ANONYMOUS, -1, 0);
   assert(ptr);
   if (trace_fp) {
-    fprintf(trace_fp, "m %lld %ld\n", (uint64_t)ptr, size);
+    fprintf(trace_fp, "m %llu %ld\n", (unsigned long long)ptr, size);
   }
   return ptr;
 }
@@ -353,7 +353,7 @@ void munmap_to_system(void *ptr, size_t size) {
   stats.munmap_size += size;
   int ret = munmap(ptr, size);
   if (trace_fp) {
-    fprintf(trace_fp, "u %lld %ld\n", (uint64_t)ptr, size);
+    fprintf(trace_fp, "u %llu %ld\n", (unsigned long long)ptr, size);
   }
   assert(ret != -1);
 }
